SCHICK: renames in seg090-seg099

+++ src/schick/datseg.h
@@ -1622,7 +1622,7 @@ extern struct struct_point g_town_positions[52];		//ds:0xa43f; seg002, seg063, s
 extern struct struct_tevent g_tevents_tab[156];		//ds:0xa66d; seg094
 extern signed char g_traveling;				//ds:0xa842; seg002, seg032, seg063, seg094
 
-extern signed short g_spell_special_aecost;	// ds:0xac0e; seg098, seg099, seg100, seg101
+extern signed int g_spell_special_aecost;	// ds:0xac0e; seg098, seg099, seg100, seg101
 
 extern int g_light_type;			// ds:0xaee8; seg101, seg107
